refactor: split TradeWidget into modals, updaters and form
