--- apps/cowswap-frontend/src/legacy/components/swap/SwapHeader/index.ts
@@ -1,3 +1,0 @@
-import SwapHeader from './SwapHeaderMod'
-export * from './SwapHeaderMod'
-export default SwapHeader
