+++ apps/cowswap-frontend/src/modules/tokensList/index.ts
@@ -1,3 +1,4 @@
 export { SelectTokenWidget } from './containers/SelectTokenWidget'
 export { ImportTokenModal } from './pure/ImportTokenModal'
 export { useOpenTokenSelectWidget } from './hooks/useOpenTokenSelectWidget'
+export { selectTokenWidgetAtom } from './state/selectTokenWidgetAtom'
