+++ apps/cowswap-frontend/src/modules/trade/containers/TradeWidget/TradeWidgetUpdaters.tsx
@@ -1,0 +1,42 @@
+import { ReactNode } from 'react'
+
+import { PriorityTokensUpdater } from '@cowprotocol/balances-and-allowances'
+import { useWalletInfo } from '@cowprotocol/wallet'
+
+import { TradeFormValidationUpdater } from 'modules/tradeFormValidation'
+import { TradeQuoteUpdater } from 'modules/tradeQuote'
+
+import { usePriorityTokenAddresses } from '../../hooks/usePriorityTokenAddresses'
+import { CommonTradeUpdater } from '../../updaters/CommonTradeUpdater'
+import { DisableNativeTokenSellingUpdater } from '../../updaters/DisableNativeTokenSellingUpdater'
+import { PriceImpactUpdater } from '../../updaters/PriceImpactUpdater'
+import { RecipientAddressUpdater } from '../../updaters/RecipientAddressUpdater'
+
+export function TradeWidgetUpdaters({
+  disableQuotePolling,
+  isExpertMode,
+  disableNativeSelling,
+  children,
+}: {
+  disableQuotePolling: boolean
+  isExpertMode: boolean
+  disableNativeSelling: boolean
+  children: ReactNode
+}) {
+  const { chainId, account } = useWalletInfo()
+  const priorityTokenAddresses = usePriorityTokenAddresses()
+
+  return (
+    <>
+      <PriorityTokensUpdater account={account} chainId={chainId} tokenAddresses={priorityTokenAddresses} />
+      <RecipientAddressUpdater />
+
+      {!disableQuotePolling && <TradeQuoteUpdater />}
+      <PriceImpactUpdater />
+      <TradeFormValidationUpdater isExpertMode={isExpertMode} />
+      <CommonTradeUpdater />
+      {disableNativeSelling && <DisableNativeTokenSellingUpdater />}
+      {children}
+    </>
+  )
+}
