+++ apps/cowswap-frontend/src/modules/trade/index.ts
@@ -1,5 +1,6 @@
 export * from './containers/TradeWidget'
 export * from './containers/TradeConfirmModal'
+export * from './containers/TradeWidgetLinks'
 export * from './pure/TradeConfirmation'
 export * from './hooks/useTradeConfirmActions'
 export * from './hooks/useTradeTypeInfo'
@@ -8,3 +9,4 @@ export * from './hooks/setupTradeState/useSetupTradeState'
 export * from './hooks/useWrapNativeFlow'
 export * from './hooks/useDerivedTradeState'
 export * from './types/TradeDerivedState'
+export * from './containers/TradeWidget/types'
