+++ apps/cowswap-frontend/src/modules/zeroApproval/containers/ZeroApprovalModal/ZeroApprovalModal.cosmos.tsx
@@ -1,13 +1,7 @@
 import { ZeroApprovalModal } from './ZeroApprovalModal'
 
-const modalState = {
-  isModalOpen: true,
-  openModal: () => {},
-  closeModal: () => {},
-}
-
 const Fixtures = {
-  default: <ZeroApprovalModal modalState={modalState} />,
+  default: <ZeroApprovalModal onDismiss={() => console.log('onDismiss')} />,
 }
 
 export default Fixtures
